+++ capi/geos_ts_c.cpp
@@ -813,10 +813,10 @@ extern "C" {
     }
 
     char
-    GEOSIdentical_r(GEOSContextHandle_t extHandle, const Geometry* g1, const Geometry* g2)
+    GEOSEqualsIdentical_r(GEOSContextHandle_t extHandle, const Geometry* g1, const Geometry* g2)
     {
         return execute(extHandle, 2, [&]() {
-            return g1->identicalTo(*g2);
+            return g1->equalsIdentical(g2);
         });
     }