+++ include/geos/geom/Point.h
@@ -143,7 +143,7 @@ public:
 
     bool equalsExact(const Geometry* other, double tolerance = 0) const override;
 
-    bool identicalTo(const Geometry& other) const override;
+    bool equalsIdentical(const Geometry* other) const override;
 
     void
     normalize(void) override